requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.19
celery==3.1.17
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7
django-cors-headers==0.13
django-extensions==1.4.6
django-redis-cache==0.13.1
django-toolbelt==0.0.1
djangorestframework==3.1.1
drf-extensions==0.2.7
gunicorn==19.2.1
hiredis==0.2.0
inflection==0.3.0
kombu==3.0.24
prettyprint==0.1.5
psycopg2==2.5.4
pytz==2015.2
redis==2.10.3
requests==2.6.0
rest-framework-ember==1.2.0
six==1.8.0
static3==0.5.1
